xrayapi/requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.4.1
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==3.0.1
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
colorama==0.4.6
comm==0.1.2
constantly==15.1.0
contourpy==1.0.7
cryptography==36.0.1
cycler==0.11.0
daphne==3.0.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.1
djangorestframework==3.14.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.4.2
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
ftfy==6.1.1
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.13.0
h5py==3.8.0
huggingface-hub==0.13.2
hyperlink==21.0.0
idna==3.3
imageio==2.27.0
incremental==21.3.0
ipykernel==6.21.1
ipython==8.10.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter-ydoc==0.2.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_fileid==0.6.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
kaggle==1.5.12
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.5
mysqlclient==2.1.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.1.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==8.0.0
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
selenium==4.1.2
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.2
stevedore==3.5.0
taming-transformers @ git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.3.21
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
torchxrayvision==1.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.27.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
uri-template==1.2.0
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.1.0
xgboost==1.7.3
y-py==0.5.5
ypy-websocket==0.8.2
zope.interface==5.4.0
